Final commit
1. AI for player and adversary
2. Let ASP runner accept two args, args[0] is port number, args[1] is "Player" or "Adversary"
3. Tested it with architecture's code
4. Use long instead of int

+++ src/edu/nyu/cs/hps/adversarialshortestpath/Helper.java
@@ -18,7 +18,7 @@ public class Helper {
         return adjacentTable;
     }
 
-    public static int countBridges(int[][] map, Set<Integer> startSet, Set<Integer> endSet){
+    public static int countBridges(long[][] map, Set<Integer> startSet, Set<Integer> endSet){
         int count = 0;
         for(Integer i : startSet){
             for(Integer j : endSet){
